+++ src/main/java/com/cse687/zirui/bookstore/order/domain/model/depreciate/DepreciationStrategy.java
@@ -1,5 +1,5 @@
 package com.cse687.zirui.bookstore.order.domain.model.depreciate;
 
 public interface DepreciationStrategy {
-    double depreciationRate();
+    float depreciationRate();
 }
